+++ Src/BNO055.c
@@ -71,9 +71,9 @@ void BNO_I2C_Configure(void){
 }
 
 void BnoUpdateEuler(BNO* bno){
-	bno->rawEulerX = ((BNO_I2C_Read(BNO055_EULER_P_MSB_ADDR) << 8) | BNO_I2C_Read(BNO055_EULER_P_LSB_ADDR))/16.0f;
-	bno->rawEulerY = ((BNO_I2C_Read(BNO055_EULER_R_MSB_ADDR) << 8) | BNO_I2C_Read(BNO055_EULER_R_LSB_ADDR))/16.0f;
-	bno->rawEulerZ = ((BNO_I2C_Read(BNO055_EULER_H_MSB_ADDR) << 8) | BNO_I2C_Read(BNO055_EULER_H_LSB_ADDR))/16.0f;
+	bno->rawEulerX = ((BNO_I2C_Read(BNO055_EULER_P_MSB_ADDR) << 8) | BNO_I2C_Read(BNO055_EULER_P_LSB_ADDR))/16;
+	bno->rawEulerY = ((BNO_I2C_Read(BNO055_EULER_R_MSB_ADDR) << 8) | BNO_I2C_Read(BNO055_EULER_R_LSB_ADDR))/16;
+	bno->rawEulerZ = ((BNO_I2C_Read(BNO055_EULER_H_MSB_ADDR) << 8) | BNO_I2C_Read(BNO055_EULER_H_LSB_ADDR))/16;
 
 	if(bno->rawEulerX > IMU_MAX_ANGLE) bno->cappedEulerX = IMU_MAX_ANGLE;
 	else if(bno->rawEulerX < -1*IMU_MAX_ANGLE) bno->cappedEulerX = -1*IMU_MAX_ANGLE;
@@ -89,5 +89,5 @@ void BnoUpdateEuler(BNO* bno){
 }
 
 void BnoPrintEuler(BNO* bno){
-	printf("X: %f\t Y: %f\t Z: %f\n",bno->rawEulerX,bno->rawEulerY,bno->rawEulerZ);
+	printf("X: %d\t Y: %d\t Z: %d\n",(int)bno->rawEulerX,(int)bno->rawEulerY,(int)bno->rawEulerZ);
 }
